Relocate *OptionsExtensions into Helpers Namespace
- Relocate the CopyOptionsExtensions, LoggingOptionsExtensions and SelectionOptionsExtensions into the Helpers namespace.

These objects are providing the methods by which PairEvaluator performs its comparisons. These are only required to be looked at by consumers if not using the PairEvaluator, or if doing manual evaluation of their own. PairEvaluator should be preferred, as its performing the robocopy-identical evaluation.
