+++ RoboSharp.Extensions/DirectoryPair.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using RoboSharp.Extensions.Helpers;
 
 namespace RoboSharp.Extensions
 {
