+++ RoboSharp.Extensions/Helpers/CopyOptionsExtensions.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace RoboSharp.Extensions
+namespace RoboSharp.Extensions.Helpers
 {
     /// <summary>
     /// Extension Methods for Copy options to assist with custom implementations
